ai-building-blocks-agent/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.19.0
azure-search-documents==11.6.0b9
backoff==2.2.1
bcrypt==4.3.0
breadability==0.1.20
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb @ git+https://github.com/chroma-core/chroma.git@c7d01a4694ca4c2ce17f4b469c88388747ca01c9
click==8.1.8
coloredlogs==15.0.1
cryptography==44.0.2
Deprecated==1.2.18
distro==1.9.0
dnacentersdk==2.8.3
docopt==0.6.2
durationpy==0.9
elastic-transport==8.17.0
elasticsearch==8.8.0
fastapi==0.115.9
fastjsonschema==2.21.1
filelock==3.18.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.3.2
google-auth==2.39.0
googleapis-common-protos==1.70.0
greenlet==3.1.1
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
isodate==0.7.2
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
keybert==0.9.0
kubernetes==32.0.1
langchain==0.3.0
langchain-core==0.3.33
langchain-text-splitters==0.3.5
langsmith==0.1.147
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
meraki==1.53.0
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
networkx==3.4.2
nltk==3.9
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.21.1
openai==0.27.0
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-instrumentation-asgi==0.53b1
opentelemetry-instrumentation-fastapi==0.53b1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
opentelemetry-util-http==0.53b1
orjson==3.10.16
overrides==7.7.0
packaging==24.2
pillow==11.2.1
pipdeptree==2.26.1
portalocker==2.10.1
posthog==3.25.0
propcache==0.2.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==4.1.0
sentencepiece==0.2.0
setuptools==79.0.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
sumy==0.11.0
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.8.0.dev20250423+cpu
torchvision==0.22.0.dev20250423+cpu
tqdm==4.67.1
transformers==4.51.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
